# Fix custom-property-pattern performance