--- a/.changeset/hot-doors-jump.md
+++ b/.changeset/hot-doors-jump.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `custom-property-pattern` performance