--- a/lib/rules/custom-property-pattern/index.js
+++ b/lib/rules/custom-property-pattern/index.js
@@ -40,8 +40,8 @@ const rule = (primary) => {
 		 */
 		function check(property) {
 			return (
-				!isStandardSyntaxProperty(property) ||
 				!isCustomProperty(property) ||
+				!isStandardSyntaxProperty(property) ||
 				regexpPattern.test(property.slice(2))
 			);
 		}
@@ -49,21 +49,23 @@ const rule = (primary) => {
 		root.walkDecls((decl) => {
 			const { prop, value } = decl;
 
-			const parsedValue = valueParser(value);
+			if (/var\(/i.test(value)) {
+				const parsedValue = valueParser(value);
 
-			parsedValue.walk((node) => {
-				if (!isValueFunction(node)) return;
+				parsedValue.walk((node) => {
+					if (!isValueFunction(node)) return;
 
-				if (node.value.toLowerCase() !== 'var') return;
+					if (node.value.toLowerCase() !== 'var') return;
 
-				const { nodes } = node;
+					const { nodes } = node;
 
-				const firstNode = nodes[0];
+					const firstNode = nodes[0];
 
-				if (!firstNode || check(firstNode.value)) return;
+					if (!firstNode || check(firstNode.value)) return;
 
-				complain(declarationValueIndex(decl) + firstNode.sourceIndex, firstNode.value, decl);
-			});
+					complain(declarationValueIndex(decl) + firstNode.sourceIndex, firstNode.value, decl);
+				});
+			}
 
 			if (check(prop)) return;