--- a/lib/rules/custom-property-pattern/index.js
+++ b/lib/rules/custom-property-pattern/index.js
@@ -20,6 +20,8 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/custom-property-pattern',
 };
 
+const VAR_FUNC_REGEX = /var\(/i;
+
 /** @type {import('stylelint').Rule} */
 const rule = (primary) => {
 	return (root, result) => {
@@ -49,7 +51,7 @@ const rule = (primary) => {
 		root.walkDecls((decl) => {
 			const { prop, value } = decl;
 
-			if (/var\(/i.test(value)) {
+			if (VAR_FUNC_REGEX.test(value)) {
 				const parsedValue = valueParser(value);
 
 				parsedValue.walk((node) => {