<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

Ref https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

Before: 
Mean: 148.86722335714288 ms
Deviation: 14.868270718260428 ms

After:
Mean: 120.50095842857144 ms
Deviation: 18.926086943302593 ms